+++ lib/src/clixon_netconf_lib.c
@@ -1356,7 +1356,7 @@ netconf_module_features(clicon_handle h)
     return retval;
 }
 
-/*! Load ietf netconf yang module and set enabled features
+/*! Load generic yang specs, ie ietf netconf yang module and set enabled features
  * @param[in] h  Clixon handle
  * @retval    0  OK
  * @retval   -1  Error
@@ -1379,6 +1379,9 @@ netconf_module_load(clicon_handle h)
     if (clicon_option_bool(h, "CLICON_XML_CHANGELOG"))
 	if (yang_spec_parse_module(h, "clixon-xml-changelog", NULL, yspec)< 0)
 	    goto done;
+    /* Load restconf yang. Note this is also a part of clixon-config */
+    if (yang_spec_parse_module(h, "clixon-restconf", NULL, yspec)< 0)
+	goto done;
     retval = 0;
  done:
     return retval;